1. Supports 1e8 intersection. 2. Fixed destroyAll issues. 3. Optimized federation timing and bufferings. 4. Batching callee streamings.

+++ arch/eggroll/roll/src/main/java/com/webank/ai/fate/eggroll/roll/api/grpc/server/RollKvServiceImpl.java
@@ -354,6 +354,10 @@ public class RollKvServiceImpl extends KVServiceGrpc.KVServiceImplBase {
                         storageMetaClient.updateFragment(fragment);
 
                         Node node = nodeIdToNode.get(fragment.getNodeId());
+                        if (node == null) {
+                            continue;
+                        }
+
                         StoreInfo storeInfoWithExactTableNameAndFragment = StoreInfo.copy(storeInfoWithExactTableName);
                         storeInfoWithExactTableNameAndFragment.setFragment(fragment.getFragmentOrder());
                         storageServiceClient.destroy(request, storeInfoWithExactTableNameAndFragment, node);
